+++ RELEASES.md
@@ -743,3 +743,13 @@ Based on:
 - [typescript v5.2.0-alpha.57] .
 ### Releases
 - [NPM v5.2.0-alpha.57] https://www.npmjs.com/package/meep-moop/v/5.2.0-alpha.57 - .
+
+## 2025-08-11 18:32:11
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.601.0 (2.680.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v5.2.0-alpha.58] .
+### Releases
+- [NPM v5.2.0-alpha.58] https://www.npmjs.com/package/meep-moop/v/5.2.0-alpha.58 - .